Refactored resources into file
