+++ lib/src/sylph_run.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 //import 'dart:io';
 
 import 'package:duration/duration.dart';
+import 'package:sylph/src/resources.dart';
 import 'package:sylph/src/validator.dart';
 import 'package:tool_base/tool_base.dart';
