+++ lib/src/validator.dart
@@ -1,5 +1,6 @@
 //import 'dart:io';
 
+import 'package:sylph/src/resources.dart';
 import 'package:tool_base/tool_base.dart';
 
 import 'bundle.dart';
